requirements.txt
python-binance
dataclasses-json